extract TCP dialer into separate module
