+++ internal/dialer/dialer_mac.go
@@ -1,4 +1,4 @@
-//go:build linux
+//go:build darwin
 
 package dialer
 
@@ -16,10 +16,10 @@ func Dialer(clientAddress *net.TCPAddr, TCPTimeout time.Duration) net.Dialer {
 		Control: func(network, address string, c syscall.RawConn) error {
 			return c.Control(func(fd uintptr) {
 				err := syscall.SetsockoptInt(int(fd), syscall.SOL_SOCKET, syscall.SO_REUSEADDR, 1)
+				if err != nil {
+					fmt.Println("设置 SO_REUSEADDR 失败:", err)
+				}
 			})
-			if err != nil {
-				fmt.Println("设置 SO_REUSEADDR 失败:", err)
-			}
 		},
 	}
 }
